Failed experiment #1: A car speed limiter that tries to limit's a car speed by modulating the gas and brake pedals
Here I tried to create a speed limiter that works by modulating throttle and brake pedals directly.
My motivation was because physics.setAITopSpeed() currently does not seem to work on AI cars when on Trackday with AI Flood enabled.

But I encountered a major issue with this speed limiter approach because of the way the API offers the pedal modulation functionality.
When setting a pedal input value, the api chooses the final value to be either your value or the original maximum the AI thinks it should be.
So because of this, I am unable to reliably limit the speed of AI cars because if the AI thinks it should be at full throttle, it will ignore my throttle reduction.
I don't understand why the API is designed this way instead of allowing me to set absolute pedal values.
So because of this limitation, I am not able to use this speed limiter approach.