Failed experiment #2: Custom AI Flood to replace the CSP AI Flood functionality to workaround the physics.setAITopSpeed() bug
Here I tried to create a custom AI flood manager that cycles AI cars around the main car and the idea is that it's used instead of the built-in CSP AI flood.
My motivation was because physics.setAITopSpeed() currently does not seem to work on AI cars when on Trackday with AI Flood enabled so I tried to roll my own flood manager.

But I encountered a major issue with this approach because after the api teleports a car, the car completely stops moving and refuses to budge for an arbitrary amount of time.
This delay in the AI regaining control of the car seems to match the timer that's used when the AI cars are in the pits in Trackday mode and they wait for some time before joining the track.

I've tried getting the cars to speed up and regaining control again after teleporting by using various physics api functions like:
- physics.setGentleStop(carIndex, false)
- physics.awakeCar(carIndex)
- physics.setAINoInput(carIndex, false, false)
- physics.setAICaution(carIndex, 0)
- physics.setCarFuel(carIndex, 100)
- physics.engageGear(carIndex, 3)
- physics.setEngineRPM(carIndex, 5000)

but the ai cars still refuse to move for a while after being teleported, and because the pedal modulation api doesn't allow me to set absolute pedal values,
I can't force the cars the let go of the brake.

Because of this limitation, I am not able to use this custom flood manager approach